refactor: alias dos módulos
