+++ src/core/types/ImageTypes.ts
